requirements.txt
# Social Media APIs
tweepy==4.14.0
praw==7.7.1
yfinance==0.2.36
requests==2.31.0

# Data Processing
scikit-learn==1.6.1
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1

# Machine Learning and NLP
tensorflow==2.16.1
transformers==4.37.2
torch==2.2.1
nltk==3.8.1
spacy==3.7.4
emoji==2.10.1
xgboost==2.0.3
prophet==1.1.5

# Visualization & Dashboard
streamlit==1.31.1
plotly==5.18.0
matplotlib==3.8.3
seaborn==0.13.2

# API Development
fastapi==0.109.2
uvicorn==0.27.1

# Testing
pytest==8.0.2

# Development Tools
jupyter==1.0.0
black==24.2.0
flake8==7.0.0 